\docType{data}
\name{censusData}
\alias{censusData}
\title{Census Bureau Database information}
\description{
  Information on all (as of Aug 2013) databases available
  for access through the Census Bureau's API at
  http://www.census.gov/developers/data/
}
\details{
  Census Bureau Datafiles
}
\keyword{datasets}

