\name{setkey}
\alias{setkey}
\title{Set the developer's key to access the API}
\usage{
  setkey(key)
}
\arguments{
  \item{key}{character string of the key you got from the
  Census Bureau.}
}
\description{
  Save the developer's key to a file for enabling use of
  the Census Bureau's API. You can request a key from
  http://api.census.gov/data/key_signup.html. You only have
  to do this once.
}

