\docType{data}
\name{censusData}
\alias{censusData}
\title{Census Bureau Database information}
\description{
  Information on all (as of Aug 2013) databases available
  for access through the Census Bureau's API at
  http://www.census.gov/developers/data/ \code{censusData}
  is a data frame consisting of the variables \itemize{
  \item \emph{ topic:} short name for the data base,
  generally one of `American Community Survey', `Census
  Summary File 1', or `Census Summary File 3' \item \emph{
  url:} url for access to the data \item \emph{year:} four
  digit year involved in data description \item
  \emph{dbname:} Census Bureau's identifier for the data
  set \item \emph{xml:} url to access meta-information,
  usually XML descriptives of all variable \item
  \emph{info:} short text with description of the data
  base. }
}
\details{
  Census Bureau Datafiles
}
\keyword{datasets}

